binarize transform should accept float threshold values

+++ src/sparsepy/cli/config_validation/saved_schemas/transform/binarize.py
@@ -57,7 +57,7 @@ class BinarizeTransformSchema(AbstractSchema):
             {
                 'name': 'binarize',
                 'params': {
-                    'binarize_threshold': int
+                    'binarize_threshold': float
                 }
             }
         )
